Fixed inconsistencies in file locations and namespaces
[~] Fixed some namespace inconsistencies in the Utilities module
[~] Created a "Scripts" subfolder under the CrossPlatform module, to be
consistent with other modules
