+++ Assets/HoloToolkit/CrossPlatform/Scripts/Reflection.meta
