+++ Assets/HoloToolkit/CrossPlatform/Scripts/SystemException.cs
