--- Assets/HoloToolkit/Input/Models/Cursor/cursor_hand_ready.fbm/templates.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: af8a6d560711c1345a4b656abd0c1ffe
-folderAsset: yes
-timeCreated: 1460418114
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
