--- Assets/HoloToolkit/Input/Models/Cursor/cursor_ready.fbm/templates.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: decdb6b28a07a534f9c82d311bb1bc69
-folderAsset: yes
-timeCreated: 1460418114
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
