+++ Assets/HoloToolkit/SpatialSound/Editor/UAudioManagerBaseEditor.cs
@@ -4,7 +4,6 @@
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
-using UnityUtilities;
 
 namespace HoloToolkit.Unity
 {
