+++ Assets/HoloToolkit/Utilities/Editor/EditorGUILayoutExtensions.cs
@@ -4,7 +4,7 @@
 using UnityEditor;
 using UnityEngine;
 
-namespace UnityUtilities
+namespace HoloToolkit.Unity
 {
     /// <summary>
     /// Extensions for the UnityEnditor.EditorGUILayout class.
