--- Assets/HoloToolkit/Utilities/Shaders/Scripts.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 84506d77f4d7258498b07c9e4863d7be
-folderAsset: yes
-timeCreated: 1457987835
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
